FITRUN-113 refactor: 컨트롤러 로직 분리
